p9iutew

+++ app/src/main/java/com/example/cobi/ais/Constants.java
@@ -11,7 +11,7 @@ public class Constants {
     static final int MIN_LSA_DISTANCE = 300;
 
     static final double MAX_SPEED = (30/3.6); // 30 km/h --> 8,33333 m/s
-    static final double DIFF_SPEED = (2);     // 2 m/s  --> 7,2 km/h
+    static final int DIFF_SPEED = (2);     // 2 m/s  --> 7,2 km/h
     static final double MIN_SPEED = (5/3.6);  // 5 km/h --> 1.39 m/s
 
     static final double MAX_ACCELERATION = 2; //m/s²
